src/main/java/com/algorithms/sorting/BubbleSort.java
package com.algorithms.sorting;

public class BubbleSort {

    public void countSwaps(int[] a) {
        int swaps = sorting(a);
        System.out.println("Array is sorted in " + swaps + " swaps.");
        System.out.println("First Element: " + a[0]);
        System.out.println("Last Element: " + a[a.length - 1]);
    }

    private int sorting(int[] a) {
        int n = a.length;
        int count = 0;
        for (int i = 0; i < n; i++) {
            for (int j = 0; j < n - 1; j++) {
                // Swap adjacent elements if they are in decreasing order
                if (a[j] > a[j + 1]) {
                    int temp = a[j + 1];
                    a[j + 1] = a[j];
                    a[j] = temp;
                    count++;
                }
            }
        }
        return count;
    }


}
